+++ openstack/obs/temporary_signedUrl.go
@@ -2,131 +2,13 @@ package obs
 
 import (
 	"errors"
-	"fmt"
 	"io"
 	"net/http"
 	"os"
 	"strings"
-	"time"
 )
 
-func (obsClient ObsClient) CreateSignedUrl(input *CreateSignedUrlInput) (output *CreateSignedUrlOutput, err error) {
-	if input == nil {
-		return nil, errors.New("CreateSignedUrlInput is nil")
-	}
-
-	params := make(map[string]string, len(input.QueryParams))
-	for key, value := range input.QueryParams {
-		params[key] = value
-	}
-
-	if input.SubResource != "" {
-		params[string(input.SubResource)] = ""
-	}
-
-	headers := make(map[string][]string, len(input.Headers))
-	for key, value := range input.Headers {
-		headers[key] = []string{value}
-	}
-
-	if input.Expires <= 0 {
-		input.Expires = 300
-	}
-
-	requestUrl, err := obsClient.doAuthTemporary(string(input.Method), input.Bucket, input.Key, params, headers, int64(input.Expires))
-	if err != nil {
-		return nil, err
-	}
-
-	output = &CreateSignedUrlOutput{
-		SignedUrl:                  requestUrl,
-		ActualSignedRequestHeaders: headers,
-	}
-	return
-}
-
-func (obsClient ObsClient) CreateBrowserBasedSignature(input *CreateBrowserBasedSignatureInput) (output *CreateBrowserBasedSignatureOutput, err error) {
-	if input == nil {
-		return nil, errors.New("CreateBrowserBasedSignatureInput is nil")
-	}
-
-	params := make(map[string]string, len(input.FormParams))
-	for key, value := range input.FormParams {
-		params[key] = value
-	}
-
-	date := time.Now().UTC()
-	shortDate := date.Format(SHORT_DATE_FORMAT)
-	longDate := date.Format(LONG_DATE_FORMAT)
-
-	credential, _ := getCredential(obsClient.conf.securityProvider.ak, obsClient.conf.region, shortDate)
-
-	if input.Expires <= 0 {
-		input.Expires = 300
-	}
-
-	expiration := date.Add(time.Second * time.Duration(input.Expires)).Format(ISO8601_DATE_FORMAT)
-	params[PARAM_ALGORITHM_AMZ_CAMEL] = V4_HASH_PREFIX
-	params[PARAM_CREDENTIAL_AMZ_CAMEL] = credential
-	params[PARAM_DATE_AMZ_CAMEL] = longDate
-
-	if obsClient.conf.securityProvider.securityToken != "" {
-		if obsClient.conf.signature == SignatureObs {
-			params[HEADER_STS_TOKEN_OBS] = obsClient.conf.securityProvider.securityToken
-		} else {
-			params[HEADER_STS_TOKEN_AMZ] = obsClient.conf.securityProvider.securityToken
-		}
-	}
-
-	matchAnyBucket := true
-	matchAnyKey := true
-	count := 5
-	if bucket := strings.TrimSpace(input.Bucket); bucket != "" {
-		params["bucket"] = bucket
-		matchAnyBucket = false
-		count--
-	}
-
-	if key := strings.TrimSpace(input.Key); key != "" {
-		params["key"] = key
-		matchAnyKey = false
-		count--
-	}
-
-	originPolicySlice := make([]string, 0, len(params)+count)
-	originPolicySlice = append(originPolicySlice, fmt.Sprintf("{\"expiration\":\"%s\",", expiration))
-	originPolicySlice = append(originPolicySlice, "\"conditions\":[")
-	for key, value := range params {
-		if _key := strings.TrimSpace(strings.ToLower(key)); _key != "" {
-			originPolicySlice = append(originPolicySlice, fmt.Sprintf("{\"%s\":\"%s\"},", _key, value))
-		}
-	}
-
-	if matchAnyBucket {
-		originPolicySlice = append(originPolicySlice, "[\"starts-with\", \"$bucket\", \"\"],")
-	}
-
-	if matchAnyKey {
-		originPolicySlice = append(originPolicySlice, "[\"starts-with\", \"$key\", \"\"],")
-	}
-
-	originPolicySlice = append(originPolicySlice, "]}")
-
-	originPolicy := strings.Join(originPolicySlice, "")
-	policy := Base64Encode([]byte(originPolicy))
-	signature := getSignature(policy, obsClient.conf.securityProvider.sk, obsClient.conf.region, shortDate)
-
-	output = &CreateBrowserBasedSignatureOutput{
-		OriginPolicy: originPolicy,
-		Policy:       policy,
-		Algorithm:    params[PARAM_ALGORITHM_AMZ_CAMEL],
-		Credential:   params[PARAM_CREDENTIAL_AMZ_CAMEL],
-		Date:         params[PARAM_DATE_AMZ_CAMEL],
-		Signature:    signature,
-	}
-	return
-}
-
+// ListBucketsWithSignedUrl lists buckets with the specified signed url and signed request headers
 func (obsClient ObsClient) ListBucketsWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *ListBucketsOutput, err error) {
 	output = &ListBucketsOutput{}
 	err = obsClient.doHttpWithSignedUrl("ListBuckets", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -136,6 +18,7 @@ func (obsClient ObsClient) ListBucketsWithSignedUrl(signedUrl string, actualSign
 	return
 }
 
+// CreateBucketWithSignedUrl creates bucket with the specified signed url and signed request headers and data
 func (obsClient ObsClient) CreateBucketWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("CreateBucket", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -145,6 +28,7 @@ func (obsClient ObsClient) CreateBucketWithSignedUrl(signedUrl string, actualSig
 	return
 }
 
+// DeleteBucketWithSignedUrl deletes bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) DeleteBucketWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("DeleteBucket", HTTP_DELETE, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -154,6 +38,7 @@ func (obsClient ObsClient) DeleteBucketWithSignedUrl(signedUrl string, actualSig
 	return
 }
 
+// SetBucketStoragePolicyWithSignedUrl sets bucket storage class with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketStoragePolicyWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketStoragePolicy", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -163,6 +48,7 @@ func (obsClient ObsClient) SetBucketStoragePolicyWithSignedUrl(signedUrl string,
 	return
 }
 
+// GetBucketStoragePolicyWithSignedUrl gets bucket storage class with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketStoragePolicyWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketStoragePolicyOutput, err error) {
 	output = &GetBucketStoragePolicyOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketStoragePolicy", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -172,6 +58,7 @@ func (obsClient ObsClient) GetBucketStoragePolicyWithSignedUrl(signedUrl string,
 	return
 }
 
+// ListObjectsWithSignedUrl lists objects in a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) ListObjectsWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *ListObjectsOutput, err error) {
 	output = &ListObjectsOutput{}
 	err = obsClient.doHttpWithSignedUrl("ListObjects", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -185,6 +72,7 @@ func (obsClient ObsClient) ListObjectsWithSignedUrl(signedUrl string, actualSign
 	return
 }
 
+// ListVersionsWithSignedUrl lists versioning objects in a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) ListVersionsWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *ListVersionsOutput, err error) {
 	output = &ListVersionsOutput{}
 	err = obsClient.doHttpWithSignedUrl("ListVersions", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -198,6 +86,8 @@ func (obsClient ObsClient) ListVersionsWithSignedUrl(signedUrl string, actualSig
 	return
 }
 
+// ListMultipartUploadsWithSignedUrl lists the multipart uploads that are initialized but not combined or aborted in a
+// specified bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) ListMultipartUploadsWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *ListMultipartUploadsOutput, err error) {
 	output = &ListMultipartUploadsOutput{}
 	err = obsClient.doHttpWithSignedUrl("ListMultipartUploads", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -207,6 +97,7 @@ func (obsClient ObsClient) ListMultipartUploadsWithSignedUrl(signedUrl string, a
 	return
 }
 
+// SetBucketQuotaWithSignedUrl sets the bucket quota with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketQuotaWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketQuota", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -216,6 +107,7 @@ func (obsClient ObsClient) SetBucketQuotaWithSignedUrl(signedUrl string, actualS
 	return
 }
 
+// GetBucketQuotaWithSignedUrl gets the bucket quota with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketQuotaWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketQuotaOutput, err error) {
 	output = &GetBucketQuotaOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketQuota", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -225,6 +117,7 @@ func (obsClient ObsClient) GetBucketQuotaWithSignedUrl(signedUrl string, actualS
 	return
 }
 
+// HeadBucketWithSignedUrl checks whether a bucket exists with the specified signed url and signed request headers
 func (obsClient ObsClient) HeadBucketWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("HeadBucket", HTTP_HEAD, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -234,6 +127,17 @@ func (obsClient ObsClient) HeadBucketWithSignedUrl(signedUrl string, actualSigne
 	return
 }
 
+// HeadObjectWithSignedUrl checks whether an object exists with the specified signed url and signed request headers
+func (obsClient ObsClient) HeadObjectWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
+	output = &BaseModel{}
+	err = obsClient.doHttpWithSignedUrl("HeadObject", HTTP_HEAD, signedUrl, actualSignedRequestHeaders, nil, output, true)
+	if err != nil {
+		output = nil
+	}
+	return
+}
+
+// GetBucketMetadataWithSignedUrl gets the metadata of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketMetadataWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketMetadataOutput, err error) {
 	output = &GetBucketMetadataOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketMetadata", HTTP_HEAD, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -245,6 +149,7 @@ func (obsClient ObsClient) GetBucketMetadataWithSignedUrl(signedUrl string, actu
 	return
 }
 
+// GetBucketStorageInfoWithSignedUrl gets storage information about a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketStorageInfoWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketStorageInfoOutput, err error) {
 	output = &GetBucketStorageInfoOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketStorageInfo", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -254,6 +159,7 @@ func (obsClient ObsClient) GetBucketStorageInfoWithSignedUrl(signedUrl string, a
 	return
 }
 
+// GetBucketLocationWithSignedUrl gets the location of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketLocationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketLocationOutput, err error) {
 	output = &GetBucketLocationOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketLocation", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -263,6 +169,7 @@ func (obsClient ObsClient) GetBucketLocationWithSignedUrl(signedUrl string, actu
 	return
 }
 
+// SetBucketAclWithSignedUrl sets the bucket ACL with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketAclWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketAcl", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -272,6 +179,7 @@ func (obsClient ObsClient) SetBucketAclWithSignedUrl(signedUrl string, actualSig
 	return
 }
 
+// GetBucketAclWithSignedUrl gets the bucket ACL with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketAclWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketAclOutput, err error) {
 	output = &GetBucketAclOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketAcl", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -281,6 +189,7 @@ func (obsClient ObsClient) GetBucketAclWithSignedUrl(signedUrl string, actualSig
 	return
 }
 
+// SetBucketPolicyWithSignedUrl sets the bucket policy with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketPolicyWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketPolicy", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -290,6 +199,7 @@ func (obsClient ObsClient) SetBucketPolicyWithSignedUrl(signedUrl string, actual
 	return
 }
 
+// GetBucketPolicyWithSignedUrl gets the bucket policy with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketPolicyWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketPolicyOutput, err error) {
 	output = &GetBucketPolicyOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketPolicy", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, false)
@@ -299,6 +209,7 @@ func (obsClient ObsClient) GetBucketPolicyWithSignedUrl(signedUrl string, actual
 	return
 }
 
+// DeleteBucketPolicyWithSignedUrl deletes the bucket policy with the specified signed url and signed request headers
 func (obsClient ObsClient) DeleteBucketPolicyWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("DeleteBucketPolicy", HTTP_DELETE, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -308,6 +219,7 @@ func (obsClient ObsClient) DeleteBucketPolicyWithSignedUrl(signedUrl string, act
 	return
 }
 
+// SetBucketCorsWithSignedUrl sets CORS rules for a bucket with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketCorsWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketCors", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -317,6 +229,7 @@ func (obsClient ObsClient) SetBucketCorsWithSignedUrl(signedUrl string, actualSi
 	return
 }
 
+// GetBucketCorsWithSignedUrl gets CORS rules of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketCorsWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketCorsOutput, err error) {
 	output = &GetBucketCorsOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketCors", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -326,6 +239,7 @@ func (obsClient ObsClient) GetBucketCorsWithSignedUrl(signedUrl string, actualSi
 	return
 }
 
+// DeleteBucketCorsWithSignedUrl deletes CORS rules of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) DeleteBucketCorsWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("DeleteBucketCors", HTTP_DELETE, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -335,6 +249,7 @@ func (obsClient ObsClient) DeleteBucketCorsWithSignedUrl(signedUrl string, actua
 	return
 }
 
+// SetBucketVersioningWithSignedUrl sets the versioning status for a bucket with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketVersioningWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketVersioning", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -344,6 +259,7 @@ func (obsClient ObsClient) SetBucketVersioningWithSignedUrl(signedUrl string, ac
 	return
 }
 
+// GetBucketVersioningWithSignedUrl gets the versioning status of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketVersioningWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketVersioningOutput, err error) {
 	output = &GetBucketVersioningOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketVersioning", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -353,6 +269,7 @@ func (obsClient ObsClient) GetBucketVersioningWithSignedUrl(signedUrl string, ac
 	return
 }
 
+// SetBucketWebsiteConfigurationWithSignedUrl sets website hosting for a bucket with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketWebsiteConfigurationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketWebsiteConfiguration", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -362,6 +279,7 @@ func (obsClient ObsClient) SetBucketWebsiteConfigurationWithSignedUrl(signedUrl 
 	return
 }
 
+// GetBucketWebsiteConfigurationWithSignedUrl gets the website hosting settings of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketWebsiteConfigurationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketWebsiteConfigurationOutput, err error) {
 	output = &GetBucketWebsiteConfigurationOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketWebsiteConfiguration", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -371,6 +289,7 @@ func (obsClient ObsClient) GetBucketWebsiteConfigurationWithSignedUrl(signedUrl 
 	return
 }
 
+// DeleteBucketWebsiteConfigurationWithSignedUrl deletes the website hosting settings of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) DeleteBucketWebsiteConfigurationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("DeleteBucketWebsiteConfiguration", HTTP_DELETE, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -380,6 +299,7 @@ func (obsClient ObsClient) DeleteBucketWebsiteConfigurationWithSignedUrl(signedU
 	return
 }
 
+// SetBucketLoggingConfigurationWithSignedUrl sets the bucket logging with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketLoggingConfigurationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketLoggingConfiguration", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -389,6 +309,7 @@ func (obsClient ObsClient) SetBucketLoggingConfigurationWithSignedUrl(signedUrl 
 	return
 }
 
+// GetBucketLoggingConfigurationWithSignedUrl gets the logging settings of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketLoggingConfigurationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketLoggingConfigurationOutput, err error) {
 	output = &GetBucketLoggingConfigurationOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketLoggingConfiguration", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -398,6 +319,7 @@ func (obsClient ObsClient) GetBucketLoggingConfigurationWithSignedUrl(signedUrl 
 	return
 }
 
+// SetBucketLifecycleConfigurationWithSignedUrl sets lifecycle rules for a bucket with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketLifecycleConfigurationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketLifecycleConfiguration", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -407,6 +329,7 @@ func (obsClient ObsClient) SetBucketLifecycleConfigurationWithSignedUrl(signedUr
 	return
 }
 
+// GetBucketLifecycleConfigurationWithSignedUrl gets lifecycle rules of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketLifecycleConfigurationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketLifecycleConfigurationOutput, err error) {
 	output = &GetBucketLifecycleConfigurationOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketLifecycleConfiguration", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -416,6 +339,7 @@ func (obsClient ObsClient) GetBucketLifecycleConfigurationWithSignedUrl(signedUr
 	return
 }
 
+// DeleteBucketLifecycleConfigurationWithSignedUrl deletes lifecycle rules of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) DeleteBucketLifecycleConfigurationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("DeleteBucketLifecycleConfiguration", HTTP_DELETE, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -425,6 +349,7 @@ func (obsClient ObsClient) DeleteBucketLifecycleConfigurationWithSignedUrl(signe
 	return
 }
 
+// SetBucketTaggingWithSignedUrl sets bucket tags with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketTaggingWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketTagging", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -434,6 +359,7 @@ func (obsClient ObsClient) SetBucketTaggingWithSignedUrl(signedUrl string, actua
 	return
 }
 
+// GetBucketTaggingWithSignedUrl gets bucket tags with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketTaggingWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketTaggingOutput, err error) {
 	output = &GetBucketTaggingOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketTagging", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -443,6 +369,7 @@ func (obsClient ObsClient) GetBucketTaggingWithSignedUrl(signedUrl string, actua
 	return
 }
 
+// DeleteBucketTaggingWithSignedUrl deletes bucket tags with the specified signed url and signed request headers
 func (obsClient ObsClient) DeleteBucketTaggingWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("DeleteBucketTagging", HTTP_DELETE, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -452,6 +379,7 @@ func (obsClient ObsClient) DeleteBucketTaggingWithSignedUrl(signedUrl string, ac
 	return
 }
 
+// SetBucketNotificationWithSignedUrl sets event notification for a bucket with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetBucketNotificationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetBucketNotification", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -461,6 +389,7 @@ func (obsClient ObsClient) SetBucketNotificationWithSignedUrl(signedUrl string, 
 	return
 }
 
+// GetBucketNotificationWithSignedUrl gets event notification settings of a bucket with the specified signed url and signed request headers
 func (obsClient ObsClient) GetBucketNotificationWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetBucketNotificationOutput, err error) {
 	output = &GetBucketNotificationOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetBucketNotification", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -470,6 +399,7 @@ func (obsClient ObsClient) GetBucketNotificationWithSignedUrl(signedUrl string, 
 	return
 }
 
+// DeleteObjectWithSignedUrl deletes an object with the specified signed url and signed request headers
 func (obsClient ObsClient) DeleteObjectWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *DeleteObjectOutput, err error) {
 	output = &DeleteObjectOutput{}
 	err = obsClient.doHttpWithSignedUrl("DeleteObject", HTTP_DELETE, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -481,6 +411,7 @@ func (obsClient ObsClient) DeleteObjectWithSignedUrl(signedUrl string, actualSig
 	return
 }
 
+// DeleteObjectsWithSignedUrl deletes objects in a batch with the specified signed url and signed request headers and data
 func (obsClient ObsClient) DeleteObjectsWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *DeleteObjectsOutput, err error) {
 	output = &DeleteObjectsOutput{}
 	err = obsClient.doHttpWithSignedUrl("DeleteObjects", HTTP_POST, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -490,6 +421,7 @@ func (obsClient ObsClient) DeleteObjectsWithSignedUrl(signedUrl string, actualSi
 	return
 }
 
+// SetObjectAclWithSignedUrl sets ACL for an object with the specified signed url and signed request headers and data
 func (obsClient ObsClient) SetObjectAclWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("SetObjectAcl", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -499,6 +431,7 @@ func (obsClient ObsClient) SetObjectAclWithSignedUrl(signedUrl string, actualSig
 	return
 }
 
+// GetObjectAclWithSignedUrl gets the ACL of an object with the specified signed url and signed request headers
 func (obsClient ObsClient) GetObjectAclWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetObjectAclOutput, err error) {
 	output = &GetObjectAclOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetObjectAcl", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -512,6 +445,7 @@ func (obsClient ObsClient) GetObjectAclWithSignedUrl(signedUrl string, actualSig
 	return
 }
 
+// RestoreObjectWithSignedUrl restores an object with the specified signed url and signed request headers and data
 func (obsClient ObsClient) RestoreObjectWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("RestoreObject", HTTP_POST, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -521,6 +455,7 @@ func (obsClient ObsClient) RestoreObjectWithSignedUrl(signedUrl string, actualSi
 	return
 }
 
+// GetObjectMetadataWithSignedUrl gets object metadata with the specified signed url and signed request headers
 func (obsClient ObsClient) GetObjectMetadataWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetObjectMetadataOutput, err error) {
 	output = &GetObjectMetadataOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetObjectMetadata", HTTP_HEAD, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -532,6 +467,7 @@ func (obsClient ObsClient) GetObjectMetadataWithSignedUrl(signedUrl string, actu
 	return
 }
 
+// GetObjectWithSignedUrl downloads object with the specified signed url and signed request headers
 func (obsClient ObsClient) GetObjectWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *GetObjectOutput, err error) {
 	output = &GetObjectOutput{}
 	err = obsClient.doHttpWithSignedUrl("GetObject", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -543,6 +479,7 @@ func (obsClient ObsClient) GetObjectWithSignedUrl(signedUrl string, actualSigned
 	return
 }
 
+// PutObjectWithSignedUrl uploads an object to the specified bucket with the specified signed url and signed request headers and data
 func (obsClient ObsClient) PutObjectWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *PutObjectOutput, err error) {
 	output = &PutObjectOutput{}
 	err = obsClient.doHttpWithSignedUrl("PutObject", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -554,6 +491,7 @@ func (obsClient ObsClient) PutObjectWithSignedUrl(signedUrl string, actualSigned
 	return
 }
 
+// PutFileWithSignedUrl uploads a file to the specified bucket with the specified signed url and signed request headers and sourceFile path
 func (obsClient ObsClient) PutFileWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, sourceFile string) (output *PutObjectOutput, err error) {
 	var data io.Reader
 	sourceFile = strings.TrimSpace(sourceFile)
@@ -562,7 +500,12 @@ func (obsClient ObsClient) PutFileWithSignedUrl(signedUrl string, actualSignedRe
 		if err != nil {
 			return nil, err
 		}
-		defer fd.Close()
+		defer func() {
+			errMsg := fd.Close()
+			if errMsg != nil {
+				doLog(LEVEL_WARN, "Failed to close file with reason: %v", errMsg)
+			}
+		}()
 
 		stat, err := fd.Stat()
 		if err != nil {
@@ -596,6 +539,7 @@ func (obsClient ObsClient) PutFileWithSignedUrl(signedUrl string, actualSignedRe
 	return
 }
 
+// CopyObjectWithSignedUrl creates a copy for an existing object with the specified signed url and signed request headers
 func (obsClient ObsClient) CopyObjectWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *CopyObjectOutput, err error) {
 	output = &CopyObjectOutput{}
 	err = obsClient.doHttpWithSignedUrl("CopyObject", HTTP_PUT, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -607,6 +551,7 @@ func (obsClient ObsClient) CopyObjectWithSignedUrl(signedUrl string, actualSigne
 	return
 }
 
+// AbortMultipartUploadWithSignedUrl aborts a multipart upload in a specified bucket by using the multipart upload ID with the specified signed url and signed request headers
 func (obsClient ObsClient) AbortMultipartUploadWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
 	output = &BaseModel{}
 	err = obsClient.doHttpWithSignedUrl("AbortMultipartUpload", HTTP_DELETE, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -616,6 +561,7 @@ func (obsClient ObsClient) AbortMultipartUploadWithSignedUrl(signedUrl string, a
 	return
 }
 
+// InitiateMultipartUploadWithSignedUrl initializes a multipart upload with the specified signed url and signed request headers
 func (obsClient ObsClient) InitiateMultipartUploadWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *InitiateMultipartUploadOutput, err error) {
 	output = &InitiateMultipartUploadOutput{}
 	err = obsClient.doHttpWithSignedUrl("InitiateMultipartUpload", HTTP_POST, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -627,6 +573,8 @@ func (obsClient ObsClient) InitiateMultipartUploadWithSignedUrl(signedUrl string
 	return
 }
 
+// UploadPartWithSignedUrl uploads a part to a specified bucket by using a specified multipart upload ID
+// with the specified signed url and signed request headers and data
 func (obsClient ObsClient) UploadPartWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *UploadPartOutput, err error) {
 	output = &UploadPartOutput{}
 	err = obsClient.doHttpWithSignedUrl("UploadPart", HTTP_PUT, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -638,6 +586,8 @@ func (obsClient ObsClient) UploadPartWithSignedUrl(signedUrl string, actualSigne
 	return
 }
 
+// CompleteMultipartUploadWithSignedUrl combines the uploaded parts in a specified bucket by using the multipart upload ID
+// with the specified signed url and signed request headers and data
 func (obsClient ObsClient) CompleteMultipartUploadWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header, data io.Reader) (output *CompleteMultipartUploadOutput, err error) {
 	output = &CompleteMultipartUploadOutput{}
 	err = obsClient.doHttpWithSignedUrl("CompleteMultipartUpload", HTTP_POST, signedUrl, actualSignedRequestHeaders, data, output, true)
@@ -649,6 +599,7 @@ func (obsClient ObsClient) CompleteMultipartUploadWithSignedUrl(signedUrl string
 	return
 }
 
+// ListPartsWithSignedUrl lists the uploaded parts in a bucket by using the multipart upload ID with the specified signed url and signed request headers
 func (obsClient ObsClient) ListPartsWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *ListPartsOutput, err error) {
 	output = &ListPartsOutput{}
 	err = obsClient.doHttpWithSignedUrl("ListParts", HTTP_GET, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -658,6 +609,7 @@ func (obsClient ObsClient) ListPartsWithSignedUrl(signedUrl string, actualSigned
 	return
 }
 
+// CopyPartWithSignedUrl copy a part to a specified bucket by using a specified multipart upload ID with the specified signed url and signed request headers
 func (obsClient ObsClient) CopyPartWithSignedUrl(signedUrl string, actualSignedRequestHeaders http.Header) (output *CopyPartOutput, err error) {
 	output = &CopyPartOutput{}
 	err = obsClient.doHttpWithSignedUrl("CopyPart", HTTP_PUT, signedUrl, actualSignedRequestHeaders, nil, output, true)
@@ -665,6 +617,36 @@ func (obsClient ObsClient) CopyPartWithSignedUrl(signedUrl string, actualSignedR
 		output = nil
 	} else {
 		ParseCopyPartOutput(output)
+	}
+	return
+}
+
+// SetBucketEncryptionWithSignedURL sets bucket encryption setting for a bucket with the specified signed url and signed request headers and data
+func (obsClient ObsClient) SetBucketEncryptionWithSignedURL(signedURL string, actualSignedRequestHeaders http.Header, data io.Reader) (output *BaseModel, err error) {
+	output = &BaseModel{}
+	err = obsClient.doHttpWithSignedUrl("SetBucketEncryption", HTTP_PUT, signedURL, actualSignedRequestHeaders, data, output, true)
+	if err != nil {
+		output = nil
+	}
+	return
+}
+
+// GetBucketEncryptionWithSignedURL gets bucket encryption setting of a bucket with the specified signed url and signed request headers
+func (obsClient ObsClient) GetBucketEncryptionWithSignedURL(signedURL string, actualSignedRequestHeaders http.Header) (output *GetBucketEncryptionOutput, err error) {
+	output = &GetBucketEncryptionOutput{}
+	err = obsClient.doHttpWithSignedUrl("GetBucketEncryption", HTTP_GET, signedURL, actualSignedRequestHeaders, nil, output, true)
+	if err != nil {
+		output = nil
+	}
+	return
+}
+
+// DeleteBucketEncryptionWithSignedURL deletes bucket encryption setting of a bucket with the specified signed url and signed request headers
+func (obsClient ObsClient) DeleteBucketEncryptionWithSignedURL(signedURL string, actualSignedRequestHeaders http.Header) (output *BaseModel, err error) {
+	output = &BaseModel{}
+	err = obsClient.doHttpWithSignedUrl("DeleteBucketEncryption", HTTP_DELETE, signedURL, actualSignedRequestHeaders, nil, output, true)
+	if err != nil {
+		output = nil
 	}
 	return
 }
